style/requirements.txt
absl-py==0.8.0
alembic==1.3.1
asgiref==3.2.3
astor==0.8.0
attrs==19.3.0
audioread==2.1.8
backcall==0.1.0
bleach==1.5.0
cachetools==3.1.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
Cython==0.29.13
decorator==4.4.0
defusedxml==0.6.0
Django==3.0.3
docopt==0.6.2
entrypoints==0.3
enum34==1.1.9
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
fluidsynth==0.2
future==0.18.2
gast==0.2.2
google-auth==1.9.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
html5lib==0.9999999
idna==2.8
image==1.5.28
imageio==2.5.0
importlib-metadata==0.23
inline==0.0.1
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.1
JPype1==0.7.0
jsonify==0.5
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
Keras==2.0.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
librosa==0.7.1
llvmlite==0.30.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
midi2audio==0.1.1
mistune==0.8.4
more-itertools==7.2.0
mysql==0.0.2
mysqlclient==1.4.6
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
notebook==6.0.1
numba==0.46.0
numpy==1.16.0
oauthlib==3.1.0
opencv-contrib-python==4.1.2.30
opencv-python==4.1.1.26
opt-einsum==3.1.0
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.1.0
pinyin==0.4.0
pipreqs==0.4.10
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyAudio==0.2.11
pycparser==2.19
pydot==1.4.1
pyFluidSynth==1.2.5
pygame==1.9.6
Pygments==2.4.2
PyMySQL==0.9.3
pyparsing==2.4.2
PyQt5==5.13.0
PyQt5-sip==4.19.18
pyqt5-tools==5.13.0.1.5
PyQtWebEngine==5.13.0
pyrsistent==0.15.4
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
PyWavelets==1.0.3
pywin32==224
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
QScintilla==2.11.2
qtconsole==4.5.5
requests==2.22.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.22
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
SoundFile==0.10.3.post1
SQLAlchemy==1.3.11
sqlparse==0.3.0
tensorboard==1.7.0
tensorflow==1.4.0

termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.39.0
traitlets==4.3.2
urllib3==1.25.7
Wave==0.0.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.6
widgetsnbextension==3.5.1
wrapt==1.11.2
xlwt==1.3.0
yarg==0.1.9
zipp==0.6.0
